java - fix apply setInterval

+++ src/src_platform_uwp/fmg/uwp/res/img/RotatedImg.cs
@@ -16,7 +16,7 @@ namespace fmg.uwp.res.img
       { }
 
       /// <summary> frequency of redrawing (in milliseconds) </summary>
-      public double RedrawInterval { get; set; } = 100;
+      public long RedrawInterval { get; set; } = 100;
 
       private DispatcherTimer _timer;
